pyproject.toml
[project]
name = "okc-py"
version = "1.3.0"
description = "Asynchronous Python wrapper for the OKC API"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.13"

authors = [
    { name = "Roman Chursanov", email = "roman@chrsnv.ru" }
]

license = { text = "MIT" }

keywords = [
    "okc",
    "okc-api",
    "async",
    "aiohttp",
    "wrapper",
    "api-client",
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP",
]

dependencies = [
    "aiohttp[speedups]>=3.13.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
]

[project.urls]
Homepage = "https://github.com/STP-Team/okc-py"
Repository = "https://github.com/STP-Team/okc-py"
Issues = "https://github.com/STP-Team/okc-py/issues"
Changelog = "https://github.com/STP-Team/okc-py/releases"

[project.optional-dependencies]
dev = [
    "ruff>=0.14.10",
    "ty>=0.0.8",
]

[build-system]
requires = ["uv-build>=0.9.21,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "okc-py",
    "python-dotenv>=1.2.1",
    "twine>=6.2.0",
]

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
# Enable common rule sets
extend-select = ["E", "F", "I", "B", "C90", "UP"]

ignore = [
    "D107", # Allow missing docstring in __init__ if trivial
    "D104", # Allow missing docstrings in package modules
]

exclude = ["build", "dist", ".venv", "__pycache__"]

[tool.ruff.lint.isort]
known-first-party = ["okc_py"]

[tool.uv.workspace]
members = [
    ".",
    ".",
    ".",
]

[tool.uv.sources]
okc-py = { workspace = true }
